Remote World List Loader
The first version of the remote world list loader.
The setup is still pretty much static, meaning that changing
anything on the UI might just break the whole script.

I'll try to make it more flexible in the next releases.

Meanwhile, this works and can be "just" dropped anywhere
on the scene.

This showcases how to setup the new VRChat String Loader
to download a remote portal list and use it.

The UI is barely decent since I HATE dealing with Unity UI.
And coupling it with UdonSharp is even worse.

I didn't want to put U# scripts on each element and then
instantiate, since I wnated to avoid having multiple scripts
hanging on the Update Loop for nothing but one click once
in a while...
But I think it's possible to call U# scripts methods even
when the Behaviour itself is "disabled"...
In which case I might just switch to use a U# script on the
panels since this is way more flexible.

Anyway, the package is VPM compatible, so you can just add
the souces folder to VCC User Packages through the Settings
menu.

Signed-off-by: Voyage <voyage@miouyouyou.fr>