Adding customisation
And avoiding any Unity call during the Error handler,
since it's called from another thread, that doesn't initialize
Unity objects, leading to crashes with even Debug.Log.

Which shows, again, that the VRChat Team didn't really test
their libraries.

Customisation has been added, by handling the display
through a WorldPanelElement.
The main RemoteWorldListLoader now feeds the data of
the current page to the WorldPanelElement, instead of
trying to manage the whole UI directly.

Note that it still manages the pagination and the size
of the page is equivalent to the size of the amount of
WorldPanelElements.

Also, synchronisation has been added since, well, you're
dealing with several PAGES of portals, so instead of asking
everyone to open the same page, the system synchronizes the
current opened portal ID.

Signed-off-by: Voyage <voyage@miouyouyou.fr>